+++ libs/react-components/src/tedi/components/form/choice-group/index.ts
@@ -1,0 +1,2 @@
+export * from './choice-group';
+export * from './choice-group.types';