+++ libs/react-components/src/tedi/components/status-badge/status-badge.stories.tsx
@@ -19,6 +19,10 @@ const meta: Meta<typeof StatusBadge> = {
     controls: {
       exclude: ['sm', 'md', 'lg', 'xl', 'xxl'],
     },
+    design: {
+      type: 'figma',
+      url: 'https://www.figma.com/file/jWiRIXhHRxwVdMSimKX2FF/TEDI-Design-System-(draft)?type=design&node-id=2385-24154&m=dev',
+    },
   },
 };